100-jump.c

+++ 0x1E-search_algorithms/100-jump.c
@@ -4,10 +4,10 @@
 #include "search_algos.h"
 
 /**
- * jump_search - Search value in a sorted array using a jump search algorithm.
- * @array: Pointer to the first element of the array to search in.
- * @size: Number of elements in the array.
- * @value: The value to search for.
+ * jump_search - Search for value in a sorted array using a jump search alg.
+ * @arr: Pointer to the first element of the array to search in.
+ * @sz: Number of elements in the array.
+ * @val: The value to search for.
  *
  * Return: The index of the value in the array, otherwise -1.
  */
@@ -21,7 +21,7 @@ int jump_search(int *arr, size_t sz, int val)
 	step = (size_t)sqrt(sz);
 	while (end < sz && arr[end] < val)
 	{
-		printf("Value checked arr[%d] = [%d]\n", (int)end, arr[end]);
+		printf("Value checked arr[%lu] = [%d]\n", end, arr[end]);
 		start = end;
 		end += step;
 
@@ -29,13 +29,13 @@ int jump_search(int *arr, size_t sz, int val)
 			return (-1);
 	}
 
-	printf("Value found between indexes [%d] and [%d]\n", (int)start, (int)end);
+	printf("Value found between indexes [%lu] and [%lu]\n", start, end);
 
 	end = (end >= sz) ? sz - 1 : end;
 
 	for (; (start <= end) && (arr[start] <= val); start++)
 	{
-		printf("Value checked arr[%d] = [%d]\n", (int)start, arr[start]);
+		printf("Value checked arr[%lu] = [%d]\n", start, arr[start]);
 		if (arr[start] == val)
 			return ((int)start);
 	}
